+++ CHANGES.rst
@@ -12,6 +12,7 @@ Unreleased
 - Updated all djangoproject.com links to reference the stable version (gh-1420)
 - Dropped support for Python 3.8, which reached end-of-life on 2024-10-07 (gh-1421)
 - Added support for Django 5.1 (gh-1388)
+- Added pagination to ``SimpleHistoryAdmin`` (gh-1277)
 
 3.7.0 (2024-05-29)
 ------------------